+++ game/pieces/king.py
@@ -1,16 +1,20 @@
 from .piece import Piece
-from exceptions import InvalidMoveKingMove
+from game.exceptions import InvalidPieceMove
 
 class King(Piece):
+
+    def __init__(self, color, board):
+        super().__init__(color, board) 
+        
     def __str__(self):
         return "♔" if self.__color__ == "WHITE" else "♚"
 
     def mov_correcto(self, from_x, from_y, to_x, to_y):
         if abs(from_x - to_x) > 1 or abs(from_y - to_y) > 1:
-            raise InvalidMoveKingMove("Movimiento no válido para el rey.")
+            raise InvalidPieceMove("Movimiento no válido para el rey.")
         
         if self.__is_blocked_by_own_piece(to_x, to_y):
-            raise InvalidMoveKingMove("Movimiento bloqueado por una pieza propia.")
+            raise InvalidPieceMove("Movimiento bloqueado por una pieza propia.")
         
         return True