+++ game/pieces/rook.py
@@ -1,26 +1,30 @@
 from .piece import Piece
-from exceptions import InvalidMoveRookMove
+from game.exceptions import InvalidPieceMove
 
 class Rook(Piece):
+    
+    def __init__(self, color, board):
+        super().__init__(color, board) 
+
     def __str__(self):
         return "♖" if self.get_color() == "WHITE" else "♜"
 
     def mov_correcto(self, from_x, from_y, to_x, to_y):
         # Verificar que el movimiento es horizontal o vertical
         if from_x != to_x and from_y != to_y:
-            raise InvalidMoveRookMove("El movimiento debe ser horizontal o vertical.")
+            raise InvalidPieceMove("El movimiento debe ser horizontal o vertical.")
 
         # Verificar si el movimiento es a la misma posición
         if from_x == to_x and from_y == to_y:
-            raise InvalidMoveRookMove("El movimiento no puede ser a la misma posición.")
+            raise InvalidPieceMove("El movimiento no puede ser a la misma posición.")
 
         # Verificar si el camino está bloqueado
         if from_x == to_x:  # Movimiento vertical
             if not self._check_path(from_x, from_y, to_y, vertical=True):
-                raise InvalidMoveRookMove("Camino bloqueado por otra pieza en movimiento vertical.")
+                raise InvalidPieceMove("Camino bloqueado por otra pieza en movimiento vertical.")
         else:  # Movimiento horizontal
             if not self._check_path(from_x, from_y, to_x, vertical=False):
-                raise InvalidMoveRookMove("Camino bloqueado por otra pieza en movimiento horizontal.")
+                raise InvalidPieceMove("Camino bloqueado por otra pieza en movimiento horizontal.")
 
         return True