+++ main.py
@@ -1,4 +1,8 @@
-from cli import main
+from game.cli import Cli
+
+def main():
+    interfaz = Cli()
+    interfaz.mostrar_menu()
 
 if __name__ == "__main__":
     main()